requirements.txt
certifi==2025.7.9
charset-normalizer==3.4.2
colorama==0.4.6
einops==0.8.1
filelock==3.16.1
fsspec==2025.3.0
huggingface-hub==0.33.4
idna==3.10
imageio==2.35.1
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==25.0
pillow==10.4.0
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.4
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
sympy==1.13.3
tifffile==2023.7.10
timm==1.0.17
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.2.3
